fix-crucible-overflow: a crucible past the int range draws its surface and reads its true volume, every total across goo types summed as a long for display while per-type storage stays int (decision diagnose-then-fix-crucible-overflow)
Cause: CrucibleBlockEntityRenderer.submitLiquidLevel summed CrucibleRenderState.poolVolume plus reservoirVolume as int. A reservoir of ROCK 1.1B and METAL 1.0B with a 100M melt pool wrapped to -2094967296, and the renderer returned before drawing at or below zero, so the crucible looked empty while it held its goo. CrucibleMelting.spawnBubblesIfGooPresent and GooBubbleParticle made the same int sum; CruciblePanelPainter's per-type total merged reservoir and pool with an int sum through GooContents.mergeWith.

Red before the fix: CrucibleHeldVolumeTest's sum read -2094967296 and its fill 0.0.

Fix: CrucibleBasin.heldVolume sums pool and reservoir as a long for the renderer, bubbles and particles; GooContents.totalVolume and GooFluidHandler.totalVolume return long; the HUD adds each type's reservoir and pool as a long; CrucibleMath drains with long arithmetic; GooFormat formats a long, its M tier open to Long.MAX_VALUE. GooContents keeps Integer per-type volumes with Codec.INT and ByteBufCodecs.VAR_INT, and GooFluidHandler stays int, because NeoForge's FluidStacksResourceHandler is int.

+++ src/main/java/com/mercuriusxeno/goo/block/crucible/CrucibleBlockEntity.java
@@ -249,7 +249,7 @@ public class CrucibleBlockEntity extends BlockEntity implements IGasketHolder, I
      *
      * @return the pool volume
      */
-    public int getPoolVolume() {
+    public long getPoolVolume() {
         return meltingItem.isEmpty() ? 0 : PartiallyMeltedItem.getContents(meltingItem).totalVolume();
     }
